third_party/blink/web_tests/external/wpt/shadow-dom/event-post-dispatch-expected.txt

This is a testharness.js-based test.
[FAIL] Event properties post dispatch when target get moved into the the shadow tree by event listener
  assert_equals: expected (object) object "[object MouseEvent]" but got (undefined) undefined
Harness: the test ran to completion.
